+++ hyperspy/external/progressbar.py
